reddit-producer/requirements.txt
praw
pyyaml
kafka-python>=2.0.2